steam_auth/requirements.txt
kivy==2.2.1
pycryptodomex==3.18.0
requests==2.31.0
python-dateutil==2.8.2
Pillow==10.0.0
certifi==2023.7.22
